Naive implementation of center, new file structure.
Implemented an algorithm for finding the center of a permutation
group using .centralizer() in a straightforward way.

Moved the DirectProduct() function to a new file,
sympy/combinatorics/group_constructs.py
